% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{aspectre}
\alias{aspectre}
\title{Aspect raster of South Tyrol reclassified to values from 1 - 9}
\format{
A raster file loaded using raster::raster()
}
\source{
\url{https://land.copernicus.eu/imagery-in-situ/eu-dem/eu-dem-v1-0-and-derived-products/eu-dem-v1.0}
}
\usage{
aspectre
}
\description{
Aspect raster for South Tyrol with a 10km buffer at 25m resolution. Calculated from dem
using the aspect function in ArcGIS Pro. Method set to planar. Original values have been
reclassified to values between 1 and 9. To see what the single values mean check out the
table 'aspect_map' in the Rgadgets package.
}
\keyword{datasets}
