% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{br_info}
\alias{br_info}
\title{Table with information about the beratungsring stations}
\format{
A data frame with 3430 rows and 7 variables:
\describe{
\item{st_id}{id of the station}
\item{st_name}{name of the station}
\item{measurement}{sensor name in german}
\item{measurement_id}{sensor id}
\item{unit}{measurement unit}
\item{column}{string needed to download the station data}
\item{factor}{numeric needed to download the station data}
}
}
\usage{
br_info
}
\description{
A data.frame containing id, name and sensors for the beratungsring stations
}
\keyword{datasets}
