% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dem}
\alias{dem}
\title{Digital Elevation Model of South Tyrol}
\format{
A raster file loaded using raster::raster()
}
\source{
\url{https://land.copernicus.eu/imagery-in-situ/eu-dem/eu-dem-v1-0-and-derived-products/eu-dem-v1.0}
}
\usage{
dem
}
\description{
Digital elevation model for South Tyrol with a 20km buffer at 25m resolution.
}
\keyword{datasets}
