% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dem100}
\alias{dem100}
\title{Digital Elevation Model of South Tyrol}
\format{
A raster file loaded using raster::raster()
}
\source{
\url{https://land.copernicus.eu/imagery-in-situ/eu-dem/eu-dem-v1-0-and-derived-products/eu-dem-v1.0}
}
\usage{
dem100
}
\description{
Digital elevation model for South Tyrol with a 10km buffer at 100m resolution. Calculated from
the dem at 25m resolution using the ArcGIS "resample" tool with bilinear interpolation.
}
\keyword{datasets}
