% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_transform.R
\name{leaflet_transform}
\alias{leaflet_transform}
\title{Transforming the spatial reference of sp or sf objects to be plotted via leaflet. Transforms the
spatial reference to the WGS84 projection (epsg:4326).}
\usage{
leaflet_transform(object)
}
\arguments{
\item{object}{Sp or sf object}
}
\description{
Transforming the spatial reference of sp or sf objects to be plotted via leaflet. Transforms the
spatial reference to the WGS84 projection (epsg:4326).
}
\examples{

points <- leaflet_transform(station_shp)
}
\keyword{leflet,}
\keyword{reference}
\keyword{spatial}
\keyword{transform,}
