% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclim.R
\name{rg_bioclim}
\alias{rg_bioclim}
\title{Bioclimatic index calculation}
\usage{
rg_bioclim(
  dt,
  index.calc = "winkler",
  frml = NULL,
  agg.func = NULL,
  period = NULL,
  set.zero = NULL,
  na.action = "keep",
  max.gap = 3,
  ...
)
}
\arguments{
\item{dt}{data.frame or data.table with a column of class 'Date' and the
columns mentioned in frml}

\item{index.calc}{string, Name of the index that you want to calculate.
See function get_index_params for further details.}

\item{frml}{A formula object that describes how to calculate the daily index values. The left side of the formula will be used to name the resulting column and the right side to calculate the daily index values.}

\item{agg.func}{A function that will be used to aggregate the daily values calculated by 'frml' for every year}

\item{period}{A sequence of numbers that indicates the day of years that should be used to aggregate the index}

\item{set.zero}{Are values below 0 possible or should they be converted to 0?}

\item{na.action}{string, what should happen with NAs in the data? If keep, NAs in the
data will translate to NAs in the output, if ignore the index is calculated by leaving
out the days with NAs and if fill the NAs can be filled up to maxgap by linear interpolation.}

\item{maxgap}{integer, up to how many consecutive missing values should be filled?}
}
\description{
This function allows you to calculate bioclimatic indices
from meteorological timeseries.
}
\examples{
rg_bioclim(dt = dt.agg,
  frml = winkler ~ T_mean - 10,
  agg.func = sum,
  period = 91:304,
  set.zero = T)
}
\keyword{bioclimatic}
\keyword{index}
