% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_beratungsring.R
\name{rg_br_get}
\alias{rg_br_get}
\title{Download meteorological data from the beratungsring stations in South Tyrol}
\source{
\url{https://github.com/GiulioGenova/SBR}
}
\usage{
rg_br_get(
  station_code,
  user,
  password,
  host,
  sensor_code = "simple",
  datestart = Sys.Date() - 1,
  dateend = Sys.Date(),
  progress = F
)
}
\arguments{
\item{station_code}{integer; Id(s) of the stations.See \link[Rgadgets]{br_info} for all available stations and their ids}

\item{user}{string, access credential}

\item{password}{string, access password}

\item{host}{string, host ip adress}

\item{sensor_code}{string; Name of the sensor(s) of interest (Temperatur 2m for 2m air temperature). See \link[Rgadgets]{br_info} for all available sensors and their names. Use 'simple' to download the most common sensors or NULL to download all sensors}

\item{datestart}{date; Starting date for the download}

\item{dateend}{date; End date for the download}

\item{progress}{display progress bar?}
}
\description{
Functions for accessing the Meteorological Data from the beratungsring stations. The function
\code{rg_br_get} can be used to download station measurements for the requested time
period and stations. Information about valid station_codes and
sensor names can be retrieved using the dataset \link[Rgadgets]{br_info}.
}
\examples{

##Not Run:
station_ids <- Rgadgets::br_info$st_id \%>\% unique()
data <- rg_br_get(station_code = station_ids, host = keyring::key_get('br_host'), password = keyring::key_get('br_password'), user = keyring::key_get('br_user'), sensor_code = NULL)

}
