% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rg_calc_save}
\alias{rg_calc_save}
\title{Adds together two raster layers, saves the result under path and returns the path as character vector}
\usage{
rg_calc_save(x, y, path, format = "GTiff", operation = "+", ...)
}
\arguments{
\item{x}{Raster layer}

\item{y}{Raster layer}

\item{path}{A character vector, the calculated Raster_layer will be stored under this path.}

\item{operation}{A character vector, one of +, -, * or / specifying how x and y should be combined}

\item{...}{Further parameters passed on to raster::writeRaster}
}
\description{
This takes two raster_layers x and y  and a character vector as filename, calculates a new raster_layer by adding raster x and raster y and then saves it to path and returns that path
}
\keyword{model,}
\keyword{predict,}
\keyword{raster}
