% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_aspect.R
\name{rg_classify_aspect}
\alias{rg_classify_aspect}
\title{Classify aspect values in degree to aspect classes like south, north, northwest...}
\usage{
rg_classify_aspect(x, method = "simple", ...)
}
\arguments{
\item{x}{vector or raster object that should be reclassified. Aspect values have to be in degree between 0 and 360}

\item{method}{string, simple for reclassification using the five classes south, north, west, east and flat or complex
to also use classes like northwest, southeast...}

\item{...}{further arguments passed on to \code{\link[raster]{reclassify}} or \code{\link[base]{cut}}}
}
\value{
reclassified dataset with the same format as x
}
\description{
Uses the table from \code{\link{aspect_map}} for complex and \code{\link{aspect_map2}} for the simple reclassification.
}
\examples{
rg_classify_aspect(0:365, method = 'complex')
}
\seealso{
\code{\link{aspect_map}}; \code{\link{aspect_map2}}
}
