% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_season.R
\name{rg_classify_season}
\alias{rg_classify_season}
\title{Classify a datetime sequence into seasons}
\usage{
rg_classify_season(timestamps, convention = "month_initials")
}
\arguments{
\item{timestamps}{date or POSIXct sequence}

\item{convention}{string, a convention that is used to determine the seasons. Can be
set to 'northern hemisphere', 'southern hemisphere' or 'month_initials'. In the last
case the initials of the months that fall into a single season will be used to name
the seasons.}
}
\description{
This function classifies single entries of a datetime sequence into the 4 seasons.
}
\examples{
classify_season(df$Datetime, convention = 'month_initials')
}
\keyword{classification}
\keyword{seasons,}
