% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R, R/load_rasters.R
\docType{data}
\name{rg_dem}
\alias{rg_dem}
\title{Digital Elevation Model of South Tyrol}
\format{
A raster file loaded using raster::raster()
}
\source{
\url{https://land.copernicus.eu/imagery-in-situ/eu-dem/eu-dem-v1-0-and-derived-products/eu-dem-v1.0}
}
\usage{
rg_dem

rg_dem()
}
\value{
a raster file
}
\description{
Digital elevation model for South Tyrol with a 10km buffer at 25m resolution.
}
\examples{
dem <- rg_dem()
}
\seealso{
\code{\link{rg_dem}}
}
\keyword{datasets}
