% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time.R
\name{rg_doyToDate}
\alias{rg_doyToDate}
\title{Transforms a day of year value to a date object}
\usage{
rg_doyToDate(doy, year)
}
\arguments{
\item{doy}{integer, day of year}

\item{year}{integer, the year for which you want the date object}
}
\value{

}
\description{
Transforms a day of year value to a date object
}
\examples{
date <- rg_doyToDate(30, year = 2018)
}
