% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enframe.R
\name{rg_enframe}
\alias{rg_enframe}
\title{Enframe a list of file paths into a tibble}
\usage{
rg_enframe(
  folder,
  pattern = ".tif$",
  recursive = F,
  sep = "_",
  sep_into = NULL
)
}
\arguments{
\item{folder}{string, path to folder where the files are stored}

\item{pattern}{string, a pattern used to identify the files, see \code{\link[base]{list.files}}}

\item{recursive}{logical, should the listing recurse into directories?}

\item{sep}{string, separator that is used to split up the basename of the files into different columns.
If character, sep is interpreted as a regular expression. The default value is a regular expression that matches any sequence of non-alphanumeric values.
If numeric, sep is interpreted as character positions to split at. Positive values start at 1 at the far-left of the string; negative value start at -1 at the far-right of the string. The length of sep should be one less than sep_into.
see \code{\link[tidyr]{separate}}}

\item{sep_into}{string, names of new variables to create as character vector. Use NA to omit the variable in the output.}
}
\value{

}
\description{
Enframe a list of file paths into a tibble
}
\examples{
folder <- 'data/files'
df <- rg_enframe(folder, pattern = '.csv$', sep = '_', sep_into = c(NA, 'year', 'type'))
}
