% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_raster.R
\name{rg_extract_raster}
\alias{rg_extract_raster}
\title{Extracting values from one or multiple raster files at point locations}
\usage{
rg_extract_raster(
  rasters,
  shp,
  id.vars,
  value.name = "value",
  variable.name = "variable",
  ...
)
}
\arguments{
\item{rasters}{An object of class RasterLayer, RasterStack or RasterBrick. Can also be a path to a raster file or multiple paths. In the last case the corresponding files will be opened in a stack.}

\item{shp}{A SpatialPointsDataFrame with the location of the points where the values should be extracted.}

\item{id.vars}{A vector with the column names of 'shp' that should be in the resulting data.frame}

\item{value.name}{Name of the column that stores the extracted values. Defaults to 'value'.}

\item{variable.name}{Nae of the column that stores the names of the input rasters. Defaults to 'variable'.}
}
\value{
A data.frame with the first column corresponding to the first column in 'shp' and two columns that contain the names of the given rasters and the extracted values, respectively.
}
\description{
This function allows you to extract the values of one or many rasters at the location of given points.
}
\examples{
extract.raster(rasters = data/dem_st.tif,
  shp = station_sp,
  id.vars = c('st_id', 'st_group'),
  value.name = value
  variable.name = variable)
}
\keyword{extraction}
\keyword{raster}
