% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{rg_fit_model}
\alias{rg_fit_model}
\title{Linear model fitting with optional best subset selection, recursive feature selection or forward future selection}
\usage{
rg_fit_model(
  predictors,
  response,
  method = "lm",
  ctrl = caret::trainControl(),
  subset = "full",
  preProc = c("center", "scale"),
  metric = ifelse(is.factor(response), "Accuracy", "RMSE"),
  maximize = ifelse(metric == "RMSE", FALSE, TRUE),
  tuneLength = 3,
  tuneGrid = NULL,
  verbose = T,
  withinSE = FALSE,
  minVar = 2,
  ...
)
}
\arguments{
\item{predictors}{a data.frame with the predictor variables as columns}

\item{response}{a vector with the response variable}

\item{method}{string, type of statistical model for more details see caret::train()}

\item{ctrl}{trainControl object}

\item{subset}{string, subset selection routine that should be carried out.
One of full, best, rfe or ffs}

\item{preProc}{string or vector, pre processing routines that should be applied
to predictors}

\item{metric}{string, performance metric that is used to select the best model,
defaults to RMSE for numeric and Accuracy for factor response}

\item{maximize}{string, should metric be maximized?}

\item{tuneLength}{numeric, see caret::train() for more details}

\item{tuneGrid}{see caret::train() for more details}

\item{withinSE}{Logical Models are only selected if they are better than the
currently best models Standard error}

\item{minVar}{Numeric. Number of variables to combine for the first selection}

\item{...}{further arguments passed on to caret::train()}
}
\description{
This function allows you to fit a model between
dependend and independend variables and calculate various selection routines.
}
\keyword{bioclimatic}
\keyword{index,}
\keyword{linear}
\keyword{model,}
\keyword{selection}
\keyword{subset}
