% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_geoms.R
\name{rg_geom_bar}
\alias{rg_geom_bar}
\title{Plot bars with cumstom design}
\usage{
rg_geom_bar(
  mapping = NULL,
  stat = "identity",
  width = 0.4,
  position = ggplot2::position_dodge(width = 0.5),
  color = "black",
  ...
)
}
\arguments{
\item{mapping}{aestetic mapping}

\item{stat}{string, how to calculate height of bars?}

\item{width}{integer, width of the bars}

\item{position}{position of the bars}

\item{color}{color of the frame around the bars}

\item{...}{parameters passed on to \code{geom_bar()}}
}
\value{

}
\description{
Wrapper around \code{geom_bar()} with default parameters to modify the design of the bars. For a description
of the function parameters see \link[ggplot2]{geom_bar}
}
\examples{
ggplot(BOD, aes(factor(Time), demand)) +
rg_geom_bar(aes(fill = factor(Time)))
}
