% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{rg_geom_equation}
\alias{rg_geom_equation}
\title{Adding model equation to ggplot. Wrapper around ggpmisc::stat_poly_eq() with standard options for label}
\usage{
rg_geom_equation(formula = y ~ x, ...)
}
\arguments{
\item{formula}{Model formula}

\item{...}{Further parameters passed on to ggpmisc::stat_poly_eq()
one of index or name.}
}
\description{
Adding model equation to ggplot. Wrapper around ggpmisc::stat_poly_eq() with standard options for label
}
\examples{

my_formula <- y~x
ggplot(data, aes(x, y)) +
geom_point() +
geom_smooth(method = 'lm', formula = my_formula) +
geom_equation(formula = my_formula)
}
\keyword{equation}
\keyword{ggplot,}
\keyword{model,}
