% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_geoms.R
\name{rg_geom_point}
\alias{rg_geom_point}
\title{Plot points with cumstom design}
\usage{
rg_geom_point(
  mapping = NULL,
  pch = 21,
  color = "white",
  size = 2,
  stroke = 1.5,
  ...
)
}
\arguments{
\item{mapping}{aestetic mapping}

\item{pch}{integer, point shape}

\item{color}{string, color of point}

\item{size}{integer, point size}

\item{stroke}{integer, width of point border}

\item{...}{parameters passed on to \code{geom_point()}}
}
\value{

}
\description{
Wrapper around \code{geom_point()} with default parameters to modify the design of the points. For a description
of the function parameters see \link[ggplot2]{geom_point}
}
\examples{
data('economics_long')
economics_long \%>\%
filter(year(date) == 2000) \%>\%
ggplot(aes(date, value)) +
geom_line() +
rg_geom_point(aes(fill = variable)) +
facet_wrap(~variable, scales = 'free') +
theme_bw()
}
