% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_params.R
\name{rg_index_params}
\alias{rg_index_params}
\title{Default index parameters for different bioclimatic indices}
\usage{
rg_index_params(x)
}
\arguments{
\item{x}{character, name of the bioclimatic index. Possible values are: 'huglin', 'winkler',
'GSTavg', 'cni', or 'rad_sum'. There are some predefined indices within this function
(Currently: Winkler, Huglin, GSTavg, rad_sum). To use this parameter,
the columns in the input df have to be
named the following way: tmax, tmin, tmean for maximum, minimum and mean
temperature and rad for solar radiation.}
}
\description{
This function stores some default properties of bioclimatic indices.
}
\examples{
get_index_params('huglin')
}
\keyword{bioclimatic}
\keyword{index,}
\keyword{parameters}
