% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging.R
\name{rg_krige}
\alias{rg_krige}
\title{Kriging of spatial data}
\usage{
rg_krige(
  frml,
  input_data,
  pred_locations,
  raster_out = F,
  dem = NULL,
  verbose = T,
  ...
)
}
\arguments{
\item{frml}{formula that defines the dependent variable as a linear model of independent variables;
suppose the dependent variable has name 'z', for ordinary and simple kriging use the formula 'z~1';
for simple kriging also define 'beta' (see below); for universal kriging,
suppose 'z' is linearly dependent on 'x' and 'y', use the formula 'z~x+y'.}

\item{input_data}{sp object containing the data to be interpolated}

\item{pred_locations}{sp object containing the prediction locations. If you want to krige values onto a continuous grid
transform a raster to point object using \code{\link[raster]{rasterToPoints}}. See example.}

\item{raster_out}{logical, indicating if an interpolated raster or a table is desired as output}

\item{dem}{raster object, a digital elevation model, will be used in the function \code{\link[raster]{rasterize}}
to transfer values from pred_locations to raster cells}

\item{...}{parameters passed on to \code{\link[automap]{autoKrige}}}
}
\value{

}
\description{
Tunction to perform kriging with automatic model selection using the function \code{\link[automap]{autoKrige}}.
Can generate continuous raster surfaces of interpolated values.
}
\examples{
dem <- raster::raster('dem.tif')
st_grd <- raster::rasterToPoints(dem, spatial = T)

data_in <- rgdal::readOGR('points.shp')   #A shapefile containing values to be interpolated in a column called 'resid'

result <- rg_krige(frml = resid ~ 1, input_data = data_in, pred_locations = st_grd, raster_out = T, dem = dem)
}
