% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelplot.R
\name{rg_levelplot}
\alias{rg_levelplot}
\title{Plot raster layers}
\usage{
rg_levelplot(
  raster,
  margin = FALSE,
  layout = NULL,
  palette = viridis::viridis(10),
  rev = F,
  classification.style = "pretty",
  nbreaks = 5,
  breaks = NULL,
  xlab = NULL,
  ylab = NULL,
  axis.labels.color = NULL,
  axis.labels.x.color = "black",
  axis.labels.y.color = "black",
  axis.labels.size = NULL,
  axis.labels.x.size = 1,
  axis.labels.y.size = 1,
  axis.labels.angle = NULL,
  axis.labels.x.angle = 0,
  axis.labels.y.angle = 0,
  axis.labels.font = NULL,
  axis.labels.x.font = "plain",
  axis.labels.y.font = "plain",
  axis.ticks = NULL,
  axis.ticks.x = F,
  axis.ticks.y = F,
  axis.line.col = "black",
  axis.line.type = 1,
  axis.line.width = 1,
  legend.labels = NULL,
  legend.label.position = NULL,
  legend.label.size = 1,
  legend.position = "left",
  legend.width = 1,
  legend.height = 1,
  legend.line.color = "black",
  legend.line.type = 1,
  legend.line.width = 1,
  legend.title = NULL,
  legend.title.size = 1,
  legend.title.font = 1,
  legend.title.col = "black",
  plot.background.col = "transparent",
  plot.title = NULL,
  plot.title.size = 1,
  plot.title.color = "black",
  strip.text = names(raster),
  strip.text.size = 1,
  strip.text.color = "black",
  strip.text.font = 1,
  strip.background.col = "transparent",
  strip.border.col = "transparent",
  ...
)
}
\arguments{
\item{raster}{raster object that should be plotted. Can be a single raster layer or a stack/brick}

\item{margin}{A list or a logical. If it is TRUE, two marginal graphics show the column (x) and row (y) summaries of the Raster* object. The summary is computed with the function mean. If it is a list, it contains parameters in ‘name=value’ form that define both margins, and may contain two other lists called ‘x’ and ‘y’ whose components affect the respective margins only.}

\item{layout}{vector to specify number of columns and rows in the form c(ncol, nrow)}

\item{palette}{string, the name of the palette or a function that returns color names such as RColorBrewer::brewer.pal(9, 'RdBu). If the name of the palette is given must be one of 'RdBu', 'BuRd', 'Magma', 'Viridis', 'Inferno'.}

\item{rev}{logical, should the palette be reversed?}

\item{classification.style}{string, style of the classification, one of "fixed", "sd", "equal", "pretty", "quantile", "kmeans", "hclust", "bclust", "fisher", "jenks", "dpih" or "headtails". See \link[classInt]{classIntervals}. Only used when breaks = NULL}

\item{nbreaks}{numeric, number of breaks. Only used when breaks = NULL}

\item{breaks}{numeric, locations at which the colors change. Do not changes the legend labels, for this use legend.labels and legend.labels.position}

\item{xlab}{string, Labels of the x axis. Supports also a vector of strings to draw multiple labels}

\item{ylab}{string, Labels of the y axis. Supports also a vector of strings to draw multiple labels}

\item{axis.labels.color}{string, color of the axis labels. Changes both x and y labels}

\item{axis.labels.x.color}{string, color of the x labels}

\item{axis.labels.y.color}{string, color of the y labels}

\item{axis.labels.size}{numeric, size of the axis labels. Changes both x and y labels}

\item{axis.labels.x.size}{numeric, size of the x labels}

\item{axis.labels.y.size}{numeric, size of the y labels}

\item{axis.labels.angle}{numeric, angle of the axis labels. Changes both x and y labels}

\item{axis.labels.x.angle}{numeric, angle of the x labels}

\item{axis.labels.y.angle}{numeric, angle of the y labels}

\item{axis.labels.font}{string, fontface of the axis labels. Changes both x and y labels}

\item{axis.labels.x.font}{string, fontface of x labels}

\item{axis.labels.y.font}{string, fontface of y labels}

\item{axis.ticks}{logical, should x and y axis ticks be displayed?}

\item{axis.ticks.x}{logical, should x axis ticks be displayed?}

\item{axis.ticks.y}{logical, should y axis ticks be displayed?}

\item{axis.line.col}{string, color of the box around the plot}

\item{axis.line.type}{numeric, linetype of the box around the plot}

\item{axis.line.width}{numeric, linewidth of the box arourn the plot}

\item{legend.labels}{string or numeric, the labels at the side of the legend. Set to legend.labels = breaks to have the same labels as breaks}

\item{legend.label.position}{numeric, the position of the legend labels along the legend. Set to legend.label.position = breaks to have the same positions as specified in breaks}

\item{legend.label.size}{numeric, size of the legend labels}

\item{legend.position}{string, position of the legend. One of 'top', 'right', 'bottom' or 'left'. Legend.title seems not to work with legend.position = 'right'}

\item{legend.width}{numeric, width of the legend}

\item{legend.height}{numeric, height of the legend}

\item{legend.line.color}{string, line color of the legend frame}

\item{legend.line.type}{numeric, linetype of the legend frame}

\item{legend.line.width}{numeric, linewidth of the legend frame}

\item{legend.title}{string, title of the legend. Seems not to work when legend.position = 'right'}

\item{legend.title.size}{= numeric, size of legend title}

\item{legend.title.font}{= numeric, font of legend title. 1 = plain, 2 = bold}

\item{legend.title.col}{= string, color of legend title}

\item{plot.background.col}{string, background color of the plot}

\item{plot.title}{string, title of the plot}

\item{plot.title.size}{numeric, size of the plot title}

\item{plot.title.color}{string, color of the plot title}

\item{strip.text}{string, vector of titles of plot strips. Length must be equal to the number of plot strips}

\item{strip.text.size}{numeric, size of strip text}

\item{strip.text.color}{string, color of strip text}

\item{strip.text.font}{numeric, font of strip text. 1 = plain, 2 = bold}

\item{strip.background.col}{string, background color of the box around strip title}

\item{strip.border.col}{string, color of the box around strip title}

\item{...}{further parameters passed on to \link[rasterVis]{levelplot-methods}}
}
\value{

}
\description{
This function is a wrapper around \link[rasterVis]{levelplot-methods} with more convenient parameter names.
}
\examples{
dem <- Rgadgets::rg_dem()

rg_levelplot(dem, palette = 'Inferno', xlab = c('X', 'test', 'test'), ylab = 'Y',
plot.background.col = 'gray70', plot.title = 'A new plot fn', plot.title.size = 3,
axis.line.col = 'black', strip.background.col = 'gray70', strip.border.col = 'red',
strip.text = c('test'), strip.text.size = 2, strip.text.color = 'white', legend.title = 'legend title',
legend.position = 'left', nbreaks = 10, axis.ticks.y = F, plot.title.color = 'red',
axis.labels.angle = 30, axis.labels.font = 'bold', axis.labels.color = 'yellow',
axis.labels.size = 3)
}
