% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_rebecka.R
\name{rg_rebecka_get}
\alias{rg_rebecka_get}
\alias{rg_rebecka_codes}
\title{Download REBECKA station data}
\source{
\url{https://api.fieldclimate.com/v1/docs/}
}
\usage{
rg_rebecka_get(
  station_code,
  publicKey = "056c0e506344aff17f6b60237d9dabcb87f1e8b4",
  privateKey,
  datestart = Sys.Date() - 1,
  dateend = Sys.Date(),
  url = "https://api.fieldclimate.com/v1",
  time_res = "raw",
  progress = T
)

rg_rebecka_codes(
  publicKey = "056c0e506344aff17f6b60237d9dabcb87f1e8b4",
  privateKey,
  url = "https://api.fieldclimate.com/v1"
)
}
\arguments{
\item{station_code}{character, IDs of the station(s) to download}

\item{publicKey}{string, public KEY used for API authentication}

\item{privateKey}{string, private KEY used for API authentication}

\item{datestart}{date, starting date for download}

\item{dateend}{date, ending date for download}

\item{url}{string, url for the API}

\item{time_res}{string, time resolution of the downloaded data, one of 'raw', 'hourly' or 'daily'.}
}
\value{

}
\description{
These functions download station data from the REBECKA stations using an API to the pessl database.
The function \code{rg_rebecka_ids} can be used to request the available station ids which are in turn
used in the function \code{rg_rebecka_get} to download data for a desired time period. Data from the
stations is available from 2017 - 2020 in 10 minute (raw), hourly and daily resolution. Downloading
requires a public and private Key. The functions internally use the package \code{reticulate} to
source a python script and connect to the API. The python script was written in python 2.7, it is
not guaranteed to work with newer python versions.
}
\section{Functions}{
\itemize{
\item \code{rg_rebecka_codes}: Request station ids for the REBECKA stations
}}

\examples{
Get a vector of available station ids
ids <- rg_rebecka_ids(publicKey = XXX, privateKey = XXX)

download data for the first station at daily resolution for one year
data <- rg_rebecka_get(station_code = ids[[1]], publicKey = XXX, privateKey = XXX, datestart = as.Date('20180101', format = '\%Y\%m\%d'), dateend = as.Date('20180105', format = '\%Y\%m\%d'), time_res = 'daily')
}
