% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rg_remove_outliers}
\alias{rg_remove_outliers}
\title{Remove outliers from a vector based on sigma test or interquartile range}
\usage{
rg_remove_outliers(
  v,
  type = "iqr",
  fill = NA,
  range = ifelse(type == "iqr", 1.5, 3),
  na.rm = FALSE
)
}
\arguments{
\item{v}{numeric vector}

\item{type}{string, one of iqr for interquartile range or sigma for sigma test}

\item{fill}{numeric or function, either a fixed value or a function such as mean from which
the replacement values can be computed}

\item{range}{numeric, this number is multiplied with the interquartile range for type = 'iqr'
or the standard deviation for type = 'sigma' and determines the width of the window in which
values are considered as NOT being outliers}

\item{na.rm}{logical; if true, any NA and NaN's are removed from v before the quantiles, means or standard deviations are computed.}
}
\description{
This function removes outliers from a vector and replaces them with an optional value based
on sigma test or the interquartile range
}
\examples{
remove_outliers(c(1, 2, 3, 100))
}
\keyword{cleaning}
\keyword{outliers,}
