% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pessl_api_R_handling.R
\name{rg_response_to_df}
\alias{rg_response_to_df}
\title{Transform response from pessl-api into a dataframe}
\usage{
rg_response_to_df(api_response, sensor_list = NULL, time_res = "daily")
}
\arguments{
\item{api_response}{list-object that is returned from the get_data function in the python script "pessl_api"}

\item{sensor_list}{a list of sensor = new_name pairs. Determines which sensors are used from the stations
and how they are named in the resulting data.frame}

\item{time_res}{string, resolution of the downloaded data. possible values are raw, hourly, daily or monthly}
}
\description{
This function allows you to transform the list-object that is returned from the pessl-api
into a data.frame
}
\keyword{api,}
\keyword{meteo}
\keyword{pessl,}
\keyword{stations}
