% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{slope}
\alias{slope}
\title{Slope raster of South Tyrol}
\format{
A raster file loaded using raster::raster()
}
\source{
\url{https://land.copernicus.eu/imagery-in-situ/eu-dem/eu-dem-v1-0-and-derived-products/eu-dem-v1.0}
}
\usage{
slope
}
\description{
Slope raster for South Tyrol with a 10km buffer at 25m resolution. Calculated from dem
using the slope function in ArcGIS Pro. Units in degree and method set to planar.
}
\keyword{datasets}
