% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_calculations.R
\name{solar_declination}
\alias{solar_declination}
\title{Calculate solar declination}
\usage{
solar_declination(doy)
}
\arguments{
\item{doy}{integer, day of year}
}
\value{
An integer or a numeric vector in radians
}
\description{
Calculates the solar declination for the given days. The angle is returned in radians, because the trigonometric R functions (sin, cos, tan) are workign in radians.
The Fourier series method is used to calculate solar declination.
}
\examples{
solar_declination(doy = 100)
}
\references{
\url{https://www.e-education.psu.edu/eme810/}
}
