% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_calculations.R
\name{solar_sunset_angle}
\alias{solar_sunset_angle}
\title{Calculate solar sunset angles for a given latitude}
\usage{
solar_sunset_angle(dec, lat = 0)
}
\arguments{
\item{dec}{numeric, solar declination values in radians. One or multiple values allowed}

\item{lat}{numeric, latitude in radians. Only a single value allowed}
}
\value{

}
\description{
Calculates solar sunset angles for a given latitude. The input values have to be in radians! Convert from degree to radians by multiplication with pi/180
As default a latitude value of 0 is used.
}
\examples{
declination <- solar_declination(1:365)
angle <- solar_sunset_angle(dec = declination, lat = 0.8138967)
}
\references{
\url{https://www.e-education.psu.edu/eme810/}
}
