% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{st_border}
\alias{st_border}
\title{Border of South Tyrol}
\format{
An sf object loaded using st_read()
}
\usage{
st_border
}
\description{
Polygon delineating the border or South Tyrol (Italy)
}
\keyword{datasets}
