% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{st_districts}
\alias{st_districts}
\title{Border of the districts in South Tyrol}
\format{
An sf object loaded using st_read()
}
\usage{
st_districts
}
\description{
Polygon delineating the border of the districts in South Tyrol (Italy)
}
\keyword{datasets}
