% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{st_municipalities}
\alias{st_municipalities}
\title{Border of Municipalities in South Tyrol}
\format{
An sf object loaded using st_read()
}
\usage{
st_municipalities
}
\description{
Polygon delineating the border of the municipalities in South Tyrol (Italy)
}
\keyword{datasets}
