% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{stations_rebecka}
\alias{stations_rebecka}
\title{Shapefile with location of rebecka meteorological stations}
\format{
An sf object loaded using st_read()
}
\usage{
stations_rebecka
}
\description{
A shapefile with point coordinates that describe the location of the rebecka meteorological stations
}
\keyword{datasets}
