% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{aspect_map}
\alias{aspect_map}
\title{Reclassify table for aspect values in degree}
\format{
A data frame with 10 rows and 4 variables:
\describe{
\item{from}{start value of aspect in degree}
\item{to}{end value of aspect in degree}
\item{code}{new integer code that has been assigned to the aspect range}
\item{aspect_str}{string representation of orientation, n = north, ne = north-east, e = east, se = south-east,
s = south, sw = south-west, w = west, nw = north-west, f = flat}
}
}
\source{
\url{https://desktop.arcgis.com/de/arcmap/10.3/tools/spatial-analyst-toolbox/how-aspect-works.htm}
}
\usage{
aspect_map
}
\description{
A data.frame containing the range of aspect values and their corresponding exposition as a string ('nort', 'sout' ...)
}
\keyword{datasets}
