% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R, R/load_rasters.R
\docType{data}
\name{rg_dem}
\alias{rg_dem}
\title{Digital Elevation Model of South Tyrol}
\format{
A raster file loaded using raster::raster()
}
\source{
\url{https://land.copernicus.eu/imagery-in-situ/eu-dem/eu-dem-v1-0-and-derived-products/eu-dem-v1.0}
}
\usage{
rg_dem

rg_dem(res = 25)
}
\arguments{
\item{res}{integer, resolution of resulting raser file in meters. Possible values include 25, 100 and 250}
}
\value{
a raster file
}
\description{
Digital elevation models for South Tyrol with resolutions at 25, 100 and 250m. Lower resolution dems
have been calculated from the dem at 25m resolution using the ArcGIS "resample" tool with bilinear interpolation.
}
\examples{
dem <- rg_dem()
}
\seealso{
\code{\link{rg_dem}}
}
\keyword{datasets}
