% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_geoms.R
\name{rg_geom_equation}
\alias{rg_geom_equation}
\title{Adding model equation to ggplot. Wrapper around \link[ggpmisc]{stat_poly_eq} with standard options for label}
\usage{
rg_geom_equation(formula, elements = c("eq", "r2", "p"), sep = ", ", ...)
}
\arguments{
\item{formula}{Formula, Model formula}

\item{elements}{String or vector of strings, The parts of the model statistics that should be included in the output}

\item{sep}{String, separator between model elements}

\item{...}{Further parameters passed on to \link[ggpmisc]{stat_poly_eq}
one of index or name.}
}
\description{
Adding model equation to ggplot. Wrapper around \link[ggpmisc]{stat_poly_eq} with standard options for label
}
\examples{


}
\keyword{equation}
\keyword{ggplot,}
\keyword{model,}
