% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_province.R
\name{rg_province_get}
\alias{rg_province_get}
\title{Bulk download Meteorological Data (v2)}
\source{
\url{https://github.com/mattia6690/MonalisR/blob/master/R/Processing.R}
}
\usage{
rg_province_get(
  dburl = NULL,
  station_code,
  sensor_code,
  datestart,
  dateend,
  format = "wide"
)
}
\arguments{
\item{dburl}{URL; URL of the Province Database. If left empty the original API will be used}

\item{station_code}{string; Code of the station ("SCODE")}

\item{sensor_code}{string; Abbreviation of the sensor of interest (e.g. "N" for Precipitation).
For possible values check \code{\link{rg_province_info}}.}

\item{datestart}{date; Starting date for the download}

\item{dateend}{date; End date for the download}

\item{format}{string, wide or long for a wide or long table as ar result}
}
\description{
Function for accessing the Meteorological Data. This function will replace the
Original Function in the next version. No save to CSV is possible anymore. On the other side
it will be possible to speed up the actual processing times and to make multiple requests at once
}
