% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{stations_matsch}
\alias{stations_matsch}
\title{Shapefile with location of matsch meteorological stations}
\format{
An sf object loaded using st_read()
}
\usage{
stations_matsch
}
\description{
A shapefile with point coordinates that describe the location of the matsch meteorological stations
}
\keyword{datasets}
