% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{stations_province}
\alias{stations_province}
\title{Shapefile with location of province meteorological stations}
\format{
An sf object loaded using st_read()
}
\usage{
stations_province
}
\description{
A shapefile with point coordinates that describe the location of the province meteorological stations
}
\keyword{datasets}
